+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.4] - 2026-01-04
+
+### Changed
+- Changed "Register For This Site" notice text to "Register for Caelis"
+- No longer hiding the register notice (now displays with updated text)
+
 ## [1.4.3] - 2026-01-04
 
 ### Changed